modularized old js code. updated README to match part 2 of bio site project
