+++ src/index.html
@@ -4,10 +4,9 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="shortcut icon" href="#">
-  <title>Test</title>
+  <title>Personal Bio Site</title>
 </head>
 <body>
-  <div id="login-form-container"></div>
-  <div id="app"></div>
+  <div id="projectsPage"></div>
 </body>
 </html>
