+++ src/javascripts/helpers/data/projects.js
@@ -1,0 +1,66 @@
+const projects = [
+  {
+    title: 'Cool Project',
+    screenshot: 'images/code_image.jpeg',
+    description: 'This is the best project', // A good project description includes 'the what', 'the why', and 'the how'.
+    technologiesUsed:
+      'HTML, CSS, Vanilla JavaScript, Version Control with Github',
+    available: true,
+    netlify: 'https://github.com/nss-evening-cohort-8/js-part-deux', // Towards the latter part of the class, you will learn how to host your projects and people will be able to view them live. Cool, right? Welp, until then, just use your GitHub link in this spot as well.
+    githubUrl: 'https://github.com/nss-evening-cohort-8/js-part-deux',
+    loom: '',
+    image: '',
+
+  },
+
+  {
+    title: 'Cool Project',
+    screenshot: 'images/code_image.jpeg',
+    description: 'This is the best project', // A good project description includes 'the what', 'the why', and 'the how'.
+    technologiesUsed:
+      'HTML, CSS, Vanilla JavaScript, Version Control with Github',
+    available: true,
+    netlify: 'https://github.com/nss-evening-cohort-8/js-part-deux', // Towards the latter part of the class, you will learn how to host your projects and people will be able to view them live. Cool, right? Welp, until then, just use your GitHub link in this spot as well.
+    githubUrl: 'https://github.com/nss-evening-cohort-8/js-part-deux',
+    loom: '',
+    image: '',
+  },
+  {
+    title: 'Cool Project',
+    screenshot: 'images/code_image.jpeg',
+    description: 'This is the best project', // A good project description includes 'the what', 'the why', and 'the how'.
+    technologiesUsed:
+      'HTML, CSS, Vanilla JavaScript, Version Control with Github',
+    available: true,
+    netlify: 'https://github.com/nss-evening-cohort-8/js-part-deux', // Towards the latter part of the class, you will learn how to host your projects and people will be able to view them live. Cool, right? Welp, until then, just use your GitHub link in this spot as well.
+    githubUrl: 'https://github.com/nss-evening-cohort-8/js-part-deux',
+    loom: '',
+    image: '',
+  },
+  {
+    title: 'Cool Project',
+    screenshot: 'images/code_image.jpeg',
+    description: 'This is the best project', // A good project description includes 'the what', 'the why', and 'the how'.
+    technologiesUsed:
+      'HTML, CSS, Vanilla JavaScript, Version Control with Github',
+    available: true,
+    netlify: 'https://github.com/nss-evening-cohort-8/js-part-deux', // Towards the latter part of the class, you will learn how to host your projects and people will be able to view them live. Cool, right? Welp, until then, just use your GitHub link in this spot as well.
+    githubUrl: 'https://github.com/nss-evening-cohort-8/js-part-deux',
+    loom: '',
+    image: '',
+  },
+  {
+    title: 'Cool Project',
+    screenshot: 'images/code_image.jpeg',
+    description: 'This is the best project', // A good project description includes 'the what', 'the why', and 'the how'.
+    technologiesUsed:
+      'HTML, CSS, Vanilla JavaScript, Version Control with Github',
+    available: true,
+    netlify: 'https://github.com/nss-evening-cohort-8/js-part-deux', // Towards the latter part of the class, you will learn how to host your projects and people will be able to view them live. Cool, right? Welp, until then, just use your GitHub link in this spot as well.
+    githubUrl: 'https://github.com/nss-evening-cohort-8/js-part-deux',
+    loom: '',
+    image: '',
+  },
+];
+
+export default projects;
